Render page in _app.jsx
