--- pages/index.js
@@ -1,8 +1,0 @@
-import React from 'react'
-import { connect } from 'react-redux'
-import { startClock, serverRenderClock } from '../store'
-import Examples from '../components/examples'
-
-const Index = () => null
-
-export default connect()(Index)
