+++ pages/onboard.jsx
@@ -1,0 +1,1 @@
+export default () => <div>Yo</div>
